Consolidate test code into DBTestCase helper
